Route if_features norm_1 through f64 to match C semantics
The IF-feature normalisation at ropus/src/dnn/lpcnet.rs:1632 computed
1/sqrt(...) entirely in f32; C reference at lpcnet_enc.c:134 uses the
double sqrt() overload from <math.h>, so 1.f / sqrt(f32_sumsq) does
the divide in double and narrows only at store. 1-ULP norm_1 drift
flowed through prod.r * norm_1 / prod.i * norm_1 and surfaced as
~1e-5 in feat 18 (dnn_pitch) of the LPCNet feature pipeline.

Diagnostic verified bit-pattern: Rust f32 path produced 0x3ed01ba5,
C double path produced 0x3ed01ba6, a side experiment routing through
f64 in Rust produced 0x3ed01ba6 — exact C match.

Switched to (1.0_f64 / ((1e-15f32 + prod.r*prod.r + prod.i*prod.i)
as f64).sqrt()) as f32 which preserves the f32 sumsq accumulation
(matching C product semantics) but routes the divide and sqrt
through f64 like libm.

Net effect on feat 18: 1.001358e-5 -> 9.894371e-6 (~1.2% reduction).
The fix is correct (matches C bit-exactly on this site) but the
rsqrt is not the dominant feat-18 contributor — residual drift now
confirmed to come from the xcorr_features cascade. Future slice.

Bounds re-locked at exact post-fix measurements per HLD "no slack":
- MAX_ABS_FEATURE_DRIFT 1.001358e-5 -> 9.894371e-6
- MAX_RMS_PER_FRAME_DRIFT 1.6713232e-6 -> 1.6514838e-6

Verified: 5/5 dred drift, 1866/1866 ropus lib, 17/17 param_grid,
clippy/fmt/build clean. No regressions.

Patch versions bumped: ropus 0.12.8 -> 0.12.9,
ropus-harness-deep-plc 0.1.3 -> 0.1.4.

+++ ropus/src/dnn/lpcnet.rs
@@ -1629,7 +1629,10 @@ impl LPCNetEncState {
         );
         for i in 1..PITCH_IF_MAX_FREQ {
             let prod = cmulc(x_cpx[i], self.prev_if[i]);
-            let norm_1 = 1.0 / (1e-15 + prod.r * prod.r + prod.i * prod.i).sqrt();
+            // Match C reference/dnn/lpcnet_enc.c:134 — sumsq stays in f32 (matches C
+            // product accumulation), divide+sqrt route through f64 like libm.
+            let norm_1 =
+                (1.0_f64 / ((1e-15f32 + prod.r * prod.r + prod.i * prod.i) as f64).sqrt()) as f32;
             let prod = Cpx {
                 r: prod.r * norm_1,
                 i: prod.i * norm_1,
